requirements.txt
python-gitlab
xlwt